Upgrade project to .net standard 1.3 (#2)
Renaming Archon.WebApi to Archon.Http. Need to do this to not cause conflicts with libraries that need to use the older version and the new version side-by-side.

A bunch of changes:

* Removing log4net integration project
We don't use it anymore and ASP.Net vnext has a built-in logging thing.

* Adding new version of AuthorizeCorrectly
It is implemented as middleware now. See
http://stackoverflow.com/questions/35656828/return-http-403-using-authorize-attribute-in-asp-net-mvc-6

* Remove DomainExceptionFilterAttribute

* Removing trailing slash & model validation attributes
Trailing slashes can now be configured globally:
http://stackoverflow.com/questions/27997814/lower-case-urls-and-trailing-slash

* Removing MethodConstrainedRouteAttribute
The new asp.net vnext routing doesn't have the issue this was solving.

* Updating AuthHeader manipulator to work with vnext as middleware

* Removing the FakeHttpHandler as we don't use it anymore
It is not as easy to test with as it claims

* Removing extension methods that no longer apply

* Removing ContextUtil which no longer applies

* Removing UriExtensions which no longer apply in .net core

* Convert CsvArrayConverter to CsvModelBinder

+semver:major